abidw: Consolidate setting options
When setting options meant to be used for the write_context, it is easy
to forget to change all relavant locations. In order to consolidate
that, introduce a set_opts function that sets various known options.

We benefit from earlier refactoring as now the write_context is passed
around to carry options for writers. Hence we can be sure, that if we
set up the context correctly (and do not use deprecated functionality),
the respective write_* function we see the options set in the context.

	* include/abg-writer.h: add set_opts(context, options) function
	* tools/abidw.cc: use the newly introduced set_opts where applicable

Signed-off-by: Matthias Maennich <maennich@google.com>