abidw: add option to omit the compilation directory
The compilation directory contains machine specific information
therefore breaks reproducibility of abidw's output across different
build paths. Hence introduce --no-comp-dir-path (as in the xml
attribute). Internally I decided to not carry on the duplication of
'dir' and 'path' and used 'comp_dir'.

Thanks to earlier changes, adding an option boils down to adding it to
set_opts and to the write_context along with some auxiliary functions
for setting and getting.

write_translation_unit uses the flag in the write_context and omits the
comp-dir-path if asked for.

	* include/abg-writer.h: add set_write_comp_dir and use it in
	set_opts
	* src/abg-writer.cc: add m_write_comp_dir and resp. getters and
	setters, (write_translation_unit) teach to respect
	write_comp_dir flag of write_context.
	* tools/abidw.cc: add option to omit the comp-dir-path when
	emitting translation_units

Signed-off-by: Matthias Maennich <maennich@google.com>