dwarf-reader: support ksymtab symbol lookup by name
The ksymtab address entry might not be actually pointing to a GLOBAL
visible symbol in .symtab of the same binary. E.g. that happens when
compiling the kernel with clang's -fsanitize-cfi* and friends. In
contrast, the ksymtab.name entry has to match an entry in .symtab. It
would otherwise upset linkers very much. So, rather than relying on the
address being resolvable in any way, we might as well look up the symbol
name in the .symtab directly as this is the symbol we later want to
analyze anyway and ksymtab.name has to match a visible symbol. In this
patch we fall back to a name lookup if anything we tried so far was not
successful. That makes the patch a bit more reviewable.

On the other hand, we should be able to entirely rely on the name based
lookup and drop the address based one. That would require some
refactoring that I did not start at this point.

I created some local test cases to validate my work. As of now,
tests/test-read-dwarf.cc has no explicit support for the
linux_kernel_mode, hence is not fully capable of asserting this
implementation.

Note, there is one additional case missing: If we have CFI binaries with
position relative relocations, we also need to touch how
populate_symbol_map_from_ksymtab_reloc works to do a similar method
there.

Long story short: if we decide to go this approach, I will refactor the
code and add tests accordingly and fill the gap for the missing cases.

Yes, this patch is a bit hacky at this point, but is supposed ot proof
that name based lookup can work here.

	* src/abg-dwarf-reader.cc (populate_symbol_map_from_ksymtab):
	add support for symtab lookup by ksymtab.name entries.
	(try_reading_first_ksymtab_entry): Likewise.

Signed-off-by: Matthias Maennich <maennich@google.com>