dwarf-reader: name based lookup: add support for aliased symbols
Symbols might be aliased in the .symtab in the sense that symbols refer
to the same address in a text section. E.g. '__memcpy' and 'memcpy'
might share the same address. When looking up the ksymtab symbol by
name, ensure we follow potential aliases to find a valid elf_symbol
object for lookup in the addr->symbol map.

Also, early exit from the lookup loops once we found a match.

	* src/abg-dwarf-reader.cc (populate_symbol_map_from_ksymtab):
        add support aliased symbols for name based symbol lookup.

Fixes: 57798599453f ("dwarf-reader: name based lookup: add support for aliased symbols")
Signed-off-by: Matthias Maennich <maennich@google.com>