corpus/writer: sort emitted corpora by path name
By sorting the corpora output, we achieve determinism for the emitted
XML file across multiple runs of abidw.

For that to happen, store a std::set of corpora (instead of
std::vector), conveniently sorted by path name. In order to restore the
property of the main corpus being the first entry, keep track of that
entry when we add the first element to the corpora set.

	* include/abg-corpus.h (corpus_group): make corpora_type a
	std::set sorted by the newly introduced SharedCorpusComparator.
	* src/abg-corpus.cc (corpus_group::priv): add main_corpus member.
	(corpus_group::add_corpus): store added corpora in the set,
	assert the actual insert and keep track of the first element as
	main_corpus.
	(corpus_group::get_main_corpus): return the stored reference to
	the main_corpus rather than assuming it being the first element.
	* src/abg-wrtier.cc (write_corpus_group): explicitly emit the
	  main_corpus element followed by the remaining ones of the set.

Signed-off-by: Matthias Maennich <maennich@google.com>