writer: completely skip over empty corpora
A corpus that has no symbols contributing to the ABI surface (e.g.
because of an exhaustive suppression), will not contribute in a later
comparison via abidiff and friends. Hence, there is no need for such
entries to appear in the ABI xml representation. This patch completely
suppresses empty corpora.

	* src/abg-writer.cc (write_corpus): completely skip empty
	corpora rather than creating an empty entry for them.

Signed-off-by: Matthias Maennich <maennich@google.com>