abg-reporter: fully qualify std::string and std::ostream
In order to not unnecessarily pollute the abigail namespace and to avoid
ambiguity at compile time, fully qualify std::string and std::string.
This also resolves the issue that include/abg-reporter.h could not be
compiled standalone and therefore was broken as a public header. It
required a `using std::string` from a prior include.

	* include/abg-reporter.h (reporter_base::report) fully qualify
	std::string and std::ostream for all overloads.
	(default_reporter::report_local_typedef_changes): Likewise.
	(default_reporter::report_local_qualified_type_changes): Likewise.
	(default_reporter::report_local_reference_type_changes): Likewise.
	(default_reporter::report_local_function_type_changes): Likewise.
	(default_reporter::report): Likewise.
	(leaf_reporter::report_changes_from_diff_maps): Likewise.
	(leaf_reporter::report): Likewise.

Suggested-by: Mark Barolak <mbar@google.com>
Signed-off-by: Matthias Maennich <maennich@google.com>