Drop traversable_base::traverse method.
The traverse method is defined in the traversable_base class but is
never used as the synonymous methods in the derived classes have
different argument types and so hide it.

The function can be removed and consequently the source file
abg-traverse.cc too.

This removes a trigger of Clang's -Werror-overloaded-virtual which is
enabled with -Wmost which is enabled by -Wall.

	* include/abg-traverse.h (class traversable_base): Add
	class-level comment about expected usage, drop declaration of
	traverse method.
	* src/Makefile.am: Remove mention of src/abg-traverse.cc.
	* src/abg-traverse.cc: Remove file as it only contains the
	unused traversable_base::traverse method.

Signed-off-by: Giuliano Procida <gprocida@google.com>
Signed-off-by: Matthias Maennich <maennich@google.com>